pac4j-springboot/src/test/java/org/pac4j/springboot/config/TwitterClientAutoConfigurationTests.java
package org.pac4j.springboot.config;

import org.junit.Test;
import org.junit.runner.RunWith;
import org.pac4j.config.client.PropertiesConstants;
import org.pac4j.oauth.client.TwitterClient;
import org.pac4j.springboot.Pac4jConfigurationProperties;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.context.properties.EnableConfigurationProperties;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.context.junit4.SpringRunner;

import static org.junit.Assert.assertNotNull;

@SpringBootTest(classes = TwitterClientAutoConfiguration.class,
    properties = {
        Pac4jConfigurationProperties.PREFIX + '.' + PropertiesConstants.TWITTER_ID + "=id",
        Pac4jConfigurationProperties.PREFIX + '.' + PropertiesConstants.TWITTER_SECRET + "=secret",
    })
@RunWith(SpringRunner.class)
@EnableConfigurationProperties(Pac4jConfigurationProperties.class)
public class TwitterClientAutoConfigurationTests {

    @Autowired
    private TwitterClient twitterClient;

    @Test
    public void verifyOperation() {
        assertNotNull(this.twitterClient);
    }
}
